股票分析/requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
pymongo==3.12.0
python-dotenv==0.19.0
Werkzeug==2.0.1
numpy==1.21.6
pandas==1.3.3
scikit-learn==0.24.2
plotly==5.3.1
click==8.0.1